Updates to InitialValues
- Deleted the Defaults.kt file, defaults very defined in multiple
places: nulls in InitialValues, default method parameters in methods in
InitialValues, and initializations where the values are used. Instead,
I moved all initial values to InitialValues. Instead of InitialValues
being nullable, a default one is always created, and is the single
source of truth for initial values.
- Removed applyStaticInitialValues since now they are applied directly.
- Moved tileSize, workerCount and highFidelityColors to InitialValues.
I think it makes sense to have all non-required parameters in the same
place, without having to define some in one way and others in another.
- Removed the initialValues variable which was being set to null after
use. Not it's a function that's being set to null, which feels cleaner
to me.
- Changed the `return this` in every InitialValues method to using
`apply`, which is more idiomatic and cleaner.
- Renamed `padding` to `preloadingPadding` to be clearer about what it
does (could be confused with `Modifier.padding()`)
- Added `rotationEnabled()`, `preloadingPadding()`,
and `bitmapFilteringEnabled()` to `InitialValues`. This should make
the `MapState` initialization cleaner, because all settings can be set
in `InitialValues`, rather than having to mutate it after creation to
continue the setup.